# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for the Our Memories journal app
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: memories
    description: Memory entries (journal entries)
  - name: stats
    description: Aggregates and dashboard data
  - name: letters
    description: Sealed letters between the two of you
  - name: auth
    description: Sign up, sign in, and current user
  - name: replies
    description: Short replies on memory entries
  - name: bucket-list
    description: Shared bucket list / wishlist
  - name: milestones
    description: Anniversaries and milestone counters
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /memories:
    get:
      operationId: listMemories
      tags: [memories]
      summary: List memories
      description: Returns all memories, newest first. Optional filters by mood, tag, year, and full-text search.
      parameters:
        - in: query
          name: search
          schema: { type: string }
          description: Free text search over title and body
        - in: query
          name: mood
          schema: { type: string }
        - in: query
          name: tag
          schema: { type: string }
        - in: query
          name: year
          schema: { type: integer }
        - in: query
          name: favorite
          schema: { type: boolean }
      responses:
        "200":
          description: A list of memories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Memory"
    post:
      operationId: createMemory
      tags: [memories]
      summary: Create a memory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMemoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Memory"

  /memories/{id}:
    get:
      operationId: getMemory
      tags: [memories]
      summary: Get a single memory by id
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: A memory
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Memory"
        "404":
          description: Not found
    patch:
      operationId: updateMemory
      tags: [memories]
      summary: Update a memory
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMemoryInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Memory"
        "404":
          description: Not found
    delete:
      operationId: deleteMemory
      tags: [memories]
      summary: Delete a memory
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted

  /memories/{id}/favorite:
    post:
      operationId: toggleFavorite
      tags: [memories]
      summary: Toggle favorite on a memory
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Memory"

  /stats/overview:
    get:
      operationId: getStatsOverview
      tags: [stats]
      summary: Headline stats for the dashboard
      responses:
        "200":
          description: Stats overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsOverview"

  /stats/mood-breakdown:
    get:
      operationId: getMoodBreakdown
      tags: [stats]
      summary: Count of memories grouped by mood
      responses:
        "200":
          description: Mood breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MoodCount"

  /stats/tags:
    get:
      operationId: getTagCloud
      tags: [stats]
      summary: All tags across memories with counts
      responses:
        "200":
          description: Tag cloud
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TagCount"

  /stats/timeline:
    get:
      operationId: getTimeline
      tags: [stats]
      summary: Memory counts grouped by month for the timeline view
      responses:
        "200":
          description: Timeline buckets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimelineBucket"

  /memories/on-this-day:
    get:
      operationId: getOnThisDay
      tags: [memories]
      summary: Memories from today's date in past years
      responses:
        "200":
          description: On this day memories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Memory"

  /memories/recent:
    get:
      operationId: getRecentMemories
      tags: [memories]
      summary: A handful of the most recent memories for the dashboard
      parameters:
        - in: query
          name: limit
          schema: { type: integer, default: 5 }
      responses:
        "200":
          description: Recent memories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Memory"

  /letters:
    get:
      operationId: listLetters
      tags: [letters]
      summary: List all letters
      description: Sealed letters return body=null until their unsealsAt date arrives.
      responses:
        "200":
          description: Letters list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Letter"
    post:
      operationId: createLetter
      tags: [letters]
      summary: Write a new sealed letter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLetterInput"
      responses:
        "201":
          description: Letter sealed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Letter"

  /letters/{id}:
    get:
      operationId: getLetter
      tags: [letters]
      summary: Open a single letter
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Letter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Letter"
        "404":
          description: Not found
    delete:
      operationId: deleteLetter
      tags: [letters]
      summary: Tear up a letter
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Removed

  /letters/{id}/open:
    post:
      operationId: markLetterRead
      tags: [letters]
      summary: Mark an unsealed letter as read
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Letter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Letter"
        "400":
          description: Letter is still sealed
        "404":
          description: Not found

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Create an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Sign in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Sign out
      responses:
        "204":
          description: Signed out

  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get the currently signed-in user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not signed in

  /memories/{id}/replies:
    get:
      operationId: listReplies
      tags: [replies]
      summary: List replies on a memory
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: A list of replies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reply"
    post:
      operationId: createReply
      tags: [replies]
      summary: Add a reply to a memory
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReplyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reply"

  /replies/{id}:
    delete:
      operationId: deleteReply
      tags: [replies]
      summary: Delete a reply
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted

  /bucket-list:
    get:
      operationId: listBucketList
      tags: [bucket-list]
      summary: List bucket list items
      responses:
        "200":
          description: A list of bucket list items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BucketListItem"
    post:
      operationId: createBucketListItem
      tags: [bucket-list]
      summary: Add an item to the bucket list
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBucketListInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BucketListItem"

  /bucket-list/{id}:
    patch:
      operationId: updateBucketListItem
      tags: [bucket-list]
      summary: Update or check off a bucket list item
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBucketListInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BucketListItem"
    delete:
      operationId: deleteBucketListItem
      tags: [bucket-list]
      summary: Remove a bucket list item
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted

  /milestones:
    get:
      operationId: listMilestones
      tags: [milestones]
      summary: List milestones
      responses:
        "200":
          description: A list of milestones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Milestone"
    post:
      operationId: createMilestone
      tags: [milestones]
      summary: Add a milestone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMilestoneInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Milestone"

  /milestones/{id}:
    patch:
      operationId: updateMilestone
      tags: [milestones]
      summary: Update a milestone
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMilestoneInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Milestone"
    delete:
      operationId: deleteMilestone
      tags: [milestones]
      summary: Remove a milestone
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    Author:
      type: string
      enum: [self, partner, both]
      description: Who wrote this memory — used for the his/hers/ours filter.

    Memory:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        body: { type: string }
        memoryDate:
          type: string
          format: date
          description: The date the memory happened (not when it was written)
        location: { type: string, nullable: true }
        mood: { type: string, nullable: true, description: "e.g. joyful, peaceful, silly, romantic, adventurous, nostalgic, grateful, bittersweet" }
        author: { $ref: "#/components/schemas/Author" }
        tags:
          type: array
          items: { type: string }
        photos:
          type: array
          items: { type: string, description: "URL to a photo" }
        favorite: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required:
        - id
        - title
        - body
        - memoryDate
        - author
        - tags
        - photos
        - favorite
        - createdAt
        - updatedAt

    CreateMemoryInput:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        body: { type: string }
        memoryDate: { type: string, format: date }
        location: { type: string, nullable: true }
        mood: { type: string, nullable: true }
        author: { $ref: "#/components/schemas/Author" }
        tags:
          type: array
          items: { type: string }
        photos:
          type: array
          items: { type: string }
      required: [title, body, memoryDate, author]

    UpdateMemoryInput:
      type: object
      properties:
        title: { type: string }
        body: { type: string }
        memoryDate: { type: string, format: date }
        location: { type: string, nullable: true }
        mood: { type: string, nullable: true }
        author: { $ref: "#/components/schemas/Author" }
        tags:
          type: array
          items: { type: string }
        photos:
          type: array
          items: { type: string }
        favorite: { type: boolean }

    StatsOverview:
      type: object
      properties:
        totalMemories: { type: integer }
        favoriteCount: { type: integer }
        photoCount: { type: integer }
        firstMemoryDate: { type: string, format: date, nullable: true }
        latestMemoryDate: { type: string, format: date, nullable: true }
        daysTogether: { type: integer, description: "Days between first and latest memory" }
        streakDays: { type: integer, description: "Consecutive days with at least one memory ending today" }
        topMood: { type: string, nullable: true }
      required: [totalMemories, favoriteCount, photoCount, daysTogether, streakDays]

    MoodCount:
      type: object
      properties:
        mood: { type: string }
        count: { type: integer }
      required: [mood, count]

    TagCount:
      type: object
      properties:
        tag: { type: string }
        count: { type: integer }
      required: [tag, count]

    TimelineBucket:
      type: object
      properties:
        month: { type: string, description: "YYYY-MM" }
        count: { type: integer }
      required: [month, count]

    LetterAuthor:
      type: string
      enum: [self, partner]
      description: Who wrote the letter.

    Letter:
      type: object
      properties:
        id: { type: string }
        fromAuthor: { $ref: "#/components/schemas/LetterAuthor" }
        subject: { type: string, nullable: true }
        body:
          type: string
          nullable: true
          description: "null when the letter is still sealed (unsealsAt is in the future)"
        unsealsAt: { type: string, format: date }
        sealed: { type: boolean, description: "True when unsealsAt is in the future" }
        read: { type: boolean }
        readAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, fromAuthor, unsealsAt, sealed, read, createdAt, updatedAt]

    CreateLetterInput:
      type: object
      properties:
        fromAuthor: { $ref: "#/components/schemas/LetterAuthor" }
        subject: { type: string, nullable: true }
        body: { type: string, minLength: 1 }
        unsealsAt: { type: string, format: date }
      required: [fromAuthor, body, unsealsAt]


    User:
      type: object
      properties:
        id: { type: string }
        email: { type: string, format: email }
        name: { type: string }
        createdAt: { type: string, format: date-time }
      required: [id, email, name, createdAt]

    RegisterInput:
      type: object
      properties:
        email: { type: string, format: email }
        name: { type: string, minLength: 1 }
        password: { type: string, minLength: 6 }
      required: [email, name, password]

    LoginInput:
      type: object
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 1 }
      required: [email, password]

    Reply:
      type: object
      properties:
        id: { type: string }
        memoryId: { type: string }
        authorId: { type: string }
        authorName: { type: string }
        body: { type: string }
        createdAt: { type: string, format: date-time }
      required: [id, memoryId, authorId, authorName, body, createdAt]

    CreateReplyInput:
      type: object
      properties:
        body: { type: string, minLength: 1 }
      required: [body]

    BucketListItem:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: string, nullable: true }
        category: { type: string, nullable: true, description: "e.g. place, restaurant, activity, experience" }
        completed: { type: boolean }
        completedAt: { type: string, format: date-time, nullable: true }
        addedById: { type: string, nullable: true }
        addedByName: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, title, completed, createdAt, updatedAt]

    CreateBucketListInput:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        description: { type: string, nullable: true }
        category: { type: string, nullable: true }
      required: [title]

    UpdateBucketListInput:
      type: object
      properties:
        title: { type: string }
        description: { type: string, nullable: true }
        category: { type: string, nullable: true }
        completed: { type: boolean }

    Milestone:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        date: { type: string, format: date }
        description: { type: string, nullable: true }
        icon: { type: string, nullable: true, description: "Emoji or short label" }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, title, date, createdAt, updatedAt]

    CreateMilestoneInput:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        date: { type: string, format: date }
        description: { type: string, nullable: true }
        icon: { type: string, nullable: true }
      required: [title, date]

    UpdateMilestoneInput:
      type: object
      properties:
        title: { type: string }
        date: { type: string, format: date }
        description: { type: string, nullable: true }
        icon: { type: string, nullable: true }