# lib/api-zod/src/generated/types/bucketListItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for the Our Memories journal app
 * OpenAPI spec version: 0.1.0
 */

export interface BucketListItem {
  id: string;
  title: string;
  description?: string | null;
  /** e.g. place, restaurant, activity, experience */
  category?: string | null;
  completed: boolean;
  completedAt?: Date | null;
  addedById?: string | null;
  addedByName?: string | null;
  createdAt: Date;
  updatedAt: Date;
}