# lib/api-zod/src/generated/types/createBucketListInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for the Our Memories journal app
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBucketListInput {
  /** @minLength 1 */
  title: string;
  description?: string | null;
  category?: string | null;
}