# lib/api-zod/src/generated/types/milestone.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for the Our Memories journal app
 * OpenAPI spec version: 0.1.0
 */

export interface Milestone {
  id: string;
  title: string;
  date: Date;
  description?: string | null;
  /** Emoji or short label */
  icon?: string | null;
  createdAt: Date;
  updatedAt: Date;
}