# lib/api-zod/src/generated/types/reply.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for the Our Memories journal app
 * OpenAPI spec version: 0.1.0
 */

export interface Reply {
  id: string;
  memoryId: string;
  authorId: string;
  authorName: string;
  body: string;
  createdAt: Date;
}